correct some comments, split lib.rs
